requirements.txt
fastapi
uvicorn
python-multipart
pillow
transformers
arabic-reshaper
python-bidi